The version than can output two record files
1. When the Captain received all the data, and each drone has received
the ACK of its own data, the time step will be recorded.
2. When the simulator run after a fixed time step (for example 1000
time steps), the status will be output. The statues contains: 1. The
amount of data that Captain has received at that time step. 2. Whether
each drone has sent its own data. If so, output 1, if not, output 0.
3. The initial position of each drone is more reasonable. The initial
distance between each drone and the Captain is identical.